requirements.txt
kivy
requests
PILLOW